[round-7] 이벤트

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -29,7 +29,7 @@ public record OrderInfo(
         return new OrderInfo(
                 order.getId(),
                 order.getUser().getUserId(),
-                order.getTotalPrice().intValue(),
+                order.getTotalPrice(),
                 order.getStatus(),
                 itemInfos,
                 order.getCouponId()
